resolved all tests
